# docs/swagger.yaml
definitions:
  domain.Client:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      photographer_id:
        type: integer
      updated_at:
        type: string
    type: object
  domain.Debt:
    properties:
      amount:
        type: integer
      client_id:
        type: integer
      occurredAt:
        type: string
    type: object
  domain.Payment:
    properties:
      amount:
        type: integer
      client_id:
        type: integer
      occurredAt:
        type: string
    type: object
  domain.Photographer:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  http_handler.AddDebtRequest:
    properties:
      amount:
        example: 500
        type: integer
      client_id:
        example: 2
        type: integer
      photographer_id:
        example: 1
        type: integer
    type: object
  http_handler.AddPaymentRequest:
    properties:
      amount:
        example: 500
        type: integer
      client_id:
        example: 2
        type: integer
      photographer_id:
        example: 1
        type: integer
    type: object
  http_handler.CreateClientRequest:
    properties:
      name:
        example: Alice
        type: string
      photographer_id:
        example: 1
        type: integer
    type: object
  http_handler.CreateClientResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  http_handler.CreatePhotographerRequest:
    properties:
      name:
        example: Alice
        type: string
    type: object
  http_handler.CreatePhotographerResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  http_handler.GetIncomesResponse:
    properties:
      payments:
        items:
          $ref: '#/definitions/domain.Payment'
        type: array
      total:
        example: 10000
        type: integer
    type: object
  http_handler.UpdateClientRequest:
    properties:
      name:
        example: Alice Updated
        type: string
    type: object
info:
  contact: {}
paths:
  /clients:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload для создания клиента
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http_handler.CreateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ID созданного клиента
          schema:
            $ref: '#/definitions/http_handler.CreateClientResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Создаёт нового клиента
      tags:
      - Clients
  /clients/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID клиента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Удаляет клиента по ID
      tags:
      - Clients
    put:
      consumes:
      - application/json
      parameters:
      - description: ID клиента
        in: path
        name: id
        required: true
        type: integer
      - description: Payload для обновления клиента
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http_handler.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Обновляет данные клиента
      tags:
      - Clients
  /clients/{photographerID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID фотографа
        in: path
        name: photographerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Client'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Возвращает список клиентов фотографа
      tags:
      - Clients
  /debt:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload для добавления задолженности
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http_handler.AddDebtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавляет задолженность для клиента фотографа
      tags:
      - Financial
  /debtors/{photographerID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID фотографа
        in: path
        name: photographerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задолженностей
          schema:
            items:
              $ref: '#/definitions/domain.Debt'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получает список должников фотографа
      tags:
      - Financial
  /incomes/{photographerID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID фотографа
        in: path
        name: photographerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список платежей и общий доход
          schema:
            $ref: '#/definitions/http_handler.GetIncomesResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получает детализированный список доходов фотографа
      tags:
      - Financial
  /payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload для добавления оплаты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http_handler.AddPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавляет оплату клиента фотографу
      tags:
      - Financial
  /photographers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Photographer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Возвращает список фотографов
      tags:
      - Photographers
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload для создания фотографа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http_handler.CreatePhotographerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ID созданного фотографа
          schema:
            $ref: '#/definitions/http_handler.CreatePhotographerResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Создаёт нового фотографа
      tags:
      - Photographers
swagger: "2.0"